// src/app/shared/interfaces/sidebar-menu.interface.ts
// export interface SidebarMenu {
//   id: number;
//   menu_th: string; // 200 Ch // ชื่อเมนูไทย
//   menu_en: string; // 200 Ch // ชื่อเมนูอังกฤษ
//   icon: string; // 200 Ch // ไอคอน
//   router: string;
//   menu_detail: string;
// }

export interface SidebarMenu {
  id: number;
  menu_th: string; // 200 Ch // ชื่อเมนูไทย
  menu_en: string; // 200 Ch // ชื่อเมนูอังกฤษ
  icon: string; // 200 Ch // ไอคอน
  router: string;
  menu_detail: string;
  code: string;
}